fix(suite): move common util to separate file to avoid cir. ESM in onboarding steps
