+++ packages/suite/src/views/onboarding/steps/SelectBackupType/FloatingSelections.tsx
@@ -22,12 +22,12 @@ import {
 } from '@trezor/theme';
 import { HELP_CENTER_MULTI_SHARE_BACKUP_URL } from '@trezor/urls';
 
-import { TrezorLink } from 'src/components/suite';
 import { Translation } from 'src/components/suite/Translation';
+import { TrezorLink } from 'src/components/suite/TrezorLink';
 
 import { LegacyOptions } from './LegacyOptions';
-import { isShamirBackupType } from './SelectBackupType';
 import { ShamirOptions } from './ShamirOptions';
+import { isShamirBackupType } from '../utils';
 
 const OptionGroupHeading = styled.div`
     display: flex;
