+++ packages/suite/src/views/onboarding/steps/utils/index.ts
@@ -1,0 +1,5 @@
+import { BackupType } from '@suite-common/suite-types';
+
+const SHAMIR_TYPES: BackupType[] = ['shamir-single', 'shamir-advanced'];
+
+export const isShamirBackupType = (type: BackupType) => SHAMIR_TYPES.includes(type);
